not able to figure out how to enforce final response in mock
